back-end/requirements.txt
Flask==2.0.1
pandas==1.3.3
scikit-learn==0.24.2
Werkzeug==2.0.1
boto3
python-dotenv